Create README.md
With the last half of my life spent across 5 different countries, I have skills in multicultural management. Along the over a decade-long career as project manager, my profile is that of a generalist with the capacity to specialise and adapt to the most complex and technical mission.
I developed expertise coordinating distributed teams. overseeing processes and communication strategies, shaping new and existing key partnerships, managing complex financial scenarios with a great variety of internal and external stakeholders as well as promptly reacting to constantly changing deadlines and conflicts. A believer in the higher purpose of education and technology, I wish to improve my footprint and my career.